Rename EditArticle, AddArticle dynamic access detail component and delete function
