+++ reactjs/trainingmanageapp/src/AddArticleComponent.js
