--- reactjs/trainingmanageapp/src/DeleteComponent.js
@@ -1,39 +1,0 @@
-import React, { Component } from 'react';
-import axios from 'axios';
-
-
-class DeleteComponent extends React.Component {
-  state = {
-    id: '',
-  }
-
-  handleChange = event => {
-    this.setState({ id: event.target.value });
-  }
-
-  handleSubmit = event => {
-    event.preventDefault();
-
-    axios.delete(`http://localhost/article/${this.state.id}`)
-      .then(res => {
-        console.log(res);
-        console.log(res.data);
-      })
-  }
-
-  render() {
-    return (
-      <div>
-        <form onSubmit={this.handleSubmit}>
-          <label>
-            Person ID:
-            <input type="text" name="id" onChange={this.handleChange} />
-          </label>
-          <button type="submit">Delete</button>
-        </form>
-      </div>
-    )
-  }
-}
-
-export default DeleteComponent
