+++ reactjs/trainingmanageapp/src/Detail/DetailComponent.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import axios from 'axios';
 import './DetailStyle.css';
-import './bootstrap.min.css';
-class DetailComponent extends React.Component{
+import '../bootstrap.min.css';
+class DetailComponent extends React.Component {
 
     constructor(props) {
         super(props);
@@ -10,7 +10,8 @@ class DetailComponent extends React.Component{
             name: '',
             email: '',
             image: '',
-            video_link: ''
+            video_link: '',
+            id: ''
         }
 
         this.handleChangeTitle = this.handleChangeTitle.bind(this);
@@ -20,12 +21,14 @@ class DetailComponent extends React.Component{
 
 
     }
+
+
     handleChangeTitle(e) {
         this.setState({
             title: e.target.value
         });
     }
-    
+
     handleChangeContent(e) {
         this.setState({
             content: e.target.value
@@ -44,77 +47,83 @@ class DetailComponent extends React.Component{
             return;
         this.createImage(files[0]);
     }
-    
-    handleChangeVideoLink(e) {
-        this.setState({
-            video_link: e.target.value
-        });
-    }
 
+    handleDeleteArticle = () => {
+     
+        axios.delete('http://localhost/article/' + this.props.match.params.id)
+        .then(res => {
+            console.log(res);
+        })
+        this.props.history.push('/');
+    }
     componentDidMount() {
-        axios.get('http://localhost/article/4')
+        //Get selected article
+        axios.get('http://localhost/article/' + this.props.match.params.id)
             .then(response => {
                 this.setState(response.data)
-                console.log(response.data);
+
             })
             .catch(function (error) {
                 console.log(error)
             });
-        axios.patch('http://localhost/article/4')
+        //Update views of article    
+        axios.patch('http://localhost/article/' + this.props.match.params.id)
             .then(response => {
-                console.log(response);
+
             })
     }
-    
-    handleChange = event => {
-        this.setState({ id: event.target.value });
-      }
-    
-    
+
+    // handleChange = event => {
+    //     this.setState({ id: event.target.value });
+    // }
+
+
     render() {
         return (
-            <div class="full">
-                
-            {/* content */}
-                <div class="container">
-                    <div class="row">
-                        <div class="col">
-                            <img src={require("./images/item-pic.jpg")} alt="..." class="img-thumbnail"/>
+            <div className="full">
+
+                {/* content */}
+                <div className="container">
+                    <div className="row">
+                        <div className="col">
+                            <img src={require("../images/item-pic.jpg")} alt="..." className="img-thumbnail" />
                         </div>
-                        <div class="col">
-                            <div class="form-group">
+                        <div className="col">
+                            <div className="form-group">
                                 <h1>Title:</h1>
-                                <p>{this.state.title}</p>               
+                                <p>{this.state.title}</p>
                             </div>
-                            <div class="form-group">
+                            <div className="form-group">
                                 <h1>Content:</h1>
                                 <p>{this.state.content}</p>
                             </div>
-                            <div class="form-group">
-                                <iframe width="500" height="280" src="https://www.youtube.com/embed/qmPmwdshCMw" 
-                                    frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture" 
+                            <div className="form-group">
+                                <iframe width="500" height="280" src="https://www.youtube.com/embed/qmPmwdshCMw"
+                                    frameborder="0" allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
                                     lowfullscreen>
                                 </iframe>
                             </div>
-                            
+
                         </div>
                     </div>
                     <br></br>
                     <br></br>
-                    <div class="row">
-                        <div class="btn-group">
-                            <button type="button" class="btn btn-primary">Edit</button>
-                            <button type="button" class="btn btn-primary">Delete</button>          
+                    <div className="row">
+                        <div className="btn-group">
+                            <button type="button" className="btn btn-primary">Edit</button>
+
+                            <button onClick={this.handleDeleteArticle} className="btn btn-primary" type="submit">Delete</button>
+
                         </div>
                     </div>
                 </div>
             </div>
         )
     }
-    
-    
-    
-    
+
+
+
+
 }
 
 export default DetailComponent
