+++ reactjs/trainingmanageapp/src/Detail/DetailStyle.css
