+++ reactjs/trainingmanageapp/src/EditArticleComponent.js
